+++ libs/frontend/domain/component/src/store/index.ts
@@ -1,4 +1,3 @@
-export * from './api.utils'
 export * from './component.model'
 export * from './component.service'
 export * from './component.service.ref'
