+++ libs/frontend/domain/renderer/src/atoms/atoms.ts
@@ -7,6 +7,7 @@ import { antdAtoms, antdPropsCustomizer } from './antd'
 import { codelabAtoms } from './codelab'
 import { htmlAtoms } from './html'
 // import { muiAtoms, muiPropsCustomizer } from './mui'
+import { nextAtoms } from './next'
 import { reactAtoms, reactPropsCustomizer } from './react'
 import type { AtomCustomizer } from './types'
 
@@ -16,6 +17,7 @@ export const allAtoms: IAtomRendererRecord = {
   ...codelabAtoms,
   ...antdAtoms,
   // ...muiAtoms,
+  ...nextAtoms,
   ...reactAtoms,
 }
 
@@ -24,6 +26,7 @@ export {
   codelabAtoms,
   htmlAtoms,
   // muiAtoms,
+  nextAtoms,
   reactAtoms,
 }
