+++ libs/frontend/domain/renderer/src/atoms/index.ts
@@ -4,5 +4,6 @@ export * from './atoms'
 export * from './codelab'
 export * from './html'
 // export * from './mui'
+export * from './next'
 export * from './react'
 export * from './types'
