+++ libs/frontend/domain/renderer/src/atoms/next/index.ts
@@ -1,0 +1,1 @@
+export * from './next-atoms'
