+++ libs/shared/utils/src/props/index.ts
@@ -1,2 +1,3 @@
+export * from './get-default-field-props'
 export * from './merge-props'
 export * from './prop-safe-stringify'
